rename hook source file
