+++ client/src/api/metadata/use-get-image.tsx
